backend/pyproject.toml
[project]
name = "accounting-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikola Nikolovski", email = "nikolovski.nikola42@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langgraph-cli (>=0.3.4,<0.4.0)",
    "langgraph (>=0.5.2,<0.6.0)",
    "langgraph-sdk (>=0.1.72,<0.2.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "google-auth (>=2.17.3,<3.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)"
]

[tool.poetry]
# This section tells Poetry where to find your packages.
packages = [{include = "accounting_agent", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"